Few changes
Now we should have the correct dataset and labels for the line chart. Also removed the dummy patient in the emotion_detect.py file, instead now we fetch the last User in teh database(In case we are not able to get the user currently logged in from the view.y file, we need to work on this problem next).
Minor changes.